+++ src/main/java/de/jatitv/wonderbagshop/defultValue/DefultValueChest_1.java
@@ -16,7 +16,7 @@ package de.jatitv.wonderbagshop.defultValue;
 public class DefultValueChest_1 {
 
     public static String Name = "§2Wonder§6Bag§7: §9small";
-    public static Integer price = 1000;
+    public static Double price = 1000.00;
 
     public static String DisplayName = "§2Wonder§6Bag §9small";
     public static Integer Item_anz = 5;
